pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-resource-allocator"
version = "0.1.0"
description = "Volunteer coordination platform for NGOs - matching volunteers to tasks."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Papan Ghosh" }]
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115",
    "jinja2>=3.1",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.5",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
# Alembic autogenerates these files from its own template - don't lint them.
"alembic/versions/*" = ["E", "F", "I", "B", "UP"]
# FastAPI's Depends() in default arguments is the idiomatic pattern, not a bug.
"app/main.py" = ["B008"]
"app/api/endpoints/*" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
